move things to common
