+++ server/src/routines/turn-action-compressor.ts
@@ -28,9 +28,10 @@ import {
 	PlayCardActionData,
 	WaitActionData,
 } from '../../../common/types/turn-action-data'
-import {PlayerSetupDefs} from '../../../common/utils/state-gen'
-import {GameController, GameControllerProps} from '../game-controller'
-import runGame from './game'
+import {PlayerSetupDefs} from 'common/game/setup-game'
+import {ServerSideGameController} from '../serverside-game-controller'
+import runGame from 'common/game/run-game'
+import {GameControllerProps} from 'common/game/game-controller'
 
 const VARIABLE_BYTE_MAX = 1 // 0xFF
 const INVALID_REPLAY = 0x00
@@ -741,7 +742,7 @@ export class TurnActionCompressor {
 	}
 
 	public async turnActionsToBuffer(
-		controller: GameController,
+		controller: ServerSideGameController,
 	): Promise<Buffer> {
 		const originalGame = controller.game as GameModel
 
@@ -749,7 +750,7 @@ export class TurnActionCompressor {
 
 		const secondPlayerSetupDefs: PlayerSetupDefs = controller.player2Defs
 
-		const newGameController = new GameController(
+		const newGameController = new ServerSideGameController(
 			firstPlayerSetupDefs,
 			secondPlayerSetupDefs,
 			{
@@ -811,7 +812,7 @@ export class TurnActionCompressor {
 				battleLog: Array<Message>
 		  }
 	> {
-		const con = new GameController(
+		const con = new ServerSideGameController(
 			firstPlayerSetupDefs,
 			secondPlayerSetupDefs,
 			{
